Add basic pre commit (#3484)
[43
PRs](https://github.com/bazelbuild/rules_rust/pulls?q=is%3Apr+typo+is%3Aclosed)
with the keyword "typo" in the title have been merged. To make the a
more scalable solution, I want to introduce pre-commit and starting
rolling out https://github.com/crate-ci/typos as a pre-commit hook.

In a follow up PR, we can also add a CI for this.