Fix 'NoneType has no attribute' error when building `rust_test` using `crate` attr with `settings:experimental_per_crate_rustc_flag` set (#3621)
When `settings:experimental_per_crate_rustc_flag` is set,
the`crate_root` parameter passed to `_collect_cfgs` is accessed. In the
case of a `rust_test` using the `crate` attribute, we were previously
passing the value of `ctx.file.crate_root` which could be `None`. This
change uses `CrateInfo.root` (which should not be `None`) of the library
target specified by the `crate` attr instead.

This raises a broader question of what it means for a `rust_test` with
`crate` to have a `crate_root` attr also specified. It looks like for
compilation purposes, this is ignored, and the `crate_root` of the
corresponding `crate` attribute is inherited regardless. Given that,
perhaps it makes sense to just `fail()` if a user specifies both `crate`
and `crate_root` attrs simultaneously?