[cargo-bazel] Cleanup MetadataGenerator usage (#3592)
The `splice` command was parsing the metadata just to serialize it back
out. Let's just write it to the filesystem directly.
Also the Generator would parse the Cargo.lock, which splice was
completely ignoring. And the vendor subcommand had also parsed the
lockfile already. All of this becomes clearer when we remove this
abstraction.