Speedup collect_deps (#3634)
The key insight is that `deps` and `proc_macro_deps` must be disjoint,
so the depset.to_list is not doing any deduping. With that out of the
way, we do a bit of code rearranging to avoid converting our lists of
deps to depsets too early; in general it is best to construct the output
depsets when creating the providers, not in the middle of the rule
implementations.

Before:
<img width="1682" height="99" alt="image"
src="https://github.com/user-attachments/assets/eb0f9080-f5d5-4ab0-bc1e-1698dfc5d888"
/>

After:
<img width="1681" height="113" alt="image"
src="https://github.com/user-attachments/assets/fc5c8344-3819-4b92-ac43-3f088a3b2f1a"
/>