feat: Extract rust_clippy_action to allow other rulesets to run clippy (#3660)
Extract the action to run clippy into an importable struct, so that
other rulesets can use it.

This is an alternative fix for
https://github.com/bazelbuild/rules_rust/issues/3607. Since `rules_lint`
does allow transitive semantics for linting, we can use the clippy
action there, and users who want clippy to be transitive can use
`rules_lint` to run it.

The tests pass, and I've also tested it manually against a branch of
`rules_lint`.

Reviewing instructions:
- Start from `rust/defs.bzl`. We expose two new functions: One to get
the crate info, and one to create the clippy action itself.
- Then, read the docstring for `run_clippy`. I've tried to select a
reasonable boundary for the arguments, but I'm definitely open to
opinions. For instance, we don't ever read `ctx.attrs` in `run_clippy`,
because we don't want to force users to implement the same arguments and
flags, but that may be wrong.
- Then, go on to the call site of `run_clippy`, to see if it preserves
the semantics you'd expect, particularly around capturing output.
- Finally, go on to the implementation of `run_clippy`