Add runtime_libs as rust compile action inputs (#3741)
The current handling of `static_runtime_lib`/`dynamic_runtime_lib` is
broken because it never adds these libs to the action. I think it
assumes they are coming from `cc_toolchain.all_files` but that's not
necessarily a safe assumption; when these libs are built from source
there's no reason to force all actions to depend on both of them via the
toolchain files if we may not end up using them. (We hit this scenario
in https://github.com/cerisier/toolchains_llvm_bootstrapped)